edited transaction

+++ ATM-BS.API/DTOS/TransactionDTO.cs
@@ -9,7 +9,7 @@ namespace ATM_BS.API.DTOS
         [RegularExpression("[0-9]{8}", ErrorMessage = "Account Number should be 8 digits long")]
         public int? ToAccountNumber { get; set; }
         [Required(ErrorMessage = "Amount cannot be empty")]
-        [Range(0, int.MaxValue, ErrorMessage = "Amount must be positive")]
+        [Range(0, double.MaxValue, ErrorMessage = "Amount must be positive")]
         public double Amount { get; set; }
     }
 }
